[HW09]Added news.html and flex
I created a new page and used the flex element to set the main five times wider than the asides. It doesn't look pretty, but I think I satisfied the homework requirements. I put the main and two aside elements as children of the div, so I hope that's ok. I got it to look like the layout you provided.